Moving spells to common code
